requirements.txt
gradio
mcp
python-dotenv
openai
aiohttp
requests
